# src/auth/dto/login-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsString } from 'class-validator';

import type { ILoginResponse } from '@otus-social/auth/interfaces/login-data.interface';

export class LoginResponseDto implements ILoginResponse {
  @ApiProperty({
    description: 'JWT access token',
    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...',
  })
  @IsString()
  @IsNotEmpty()
  public access_token: string;
}